requirements.txt
tkinterdnd2==0.3.0
Pillow==8.3.2
